requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic[email]==2.4.2
deepface==0.0.79
numpy==1.26.2
python-dotenv==1.0.0
pillow==10.1.0
tensorflow==2.15.0
opencv-python==4.8.1.78
pymysql==1.1.0
alembic==1.12.1
pytest==7.4.3
httpx==0.25.1
loguru==0.7.2
fastapi-pagination==0.12.12
python-decouple==3.8
redis==5.0.1
celery==5.3.4
